moving specs to their own folders
